Fix a find/replace infinite loop which manifested itself on 64 bit builds [Andres Freund]
git-svn-id: svn://svn.pgadmin.org/trunk@7941 a7884b65-44f6-0310-8a51-81a127f17b15

+++ pgadmin/ctl/ctlSQLBox.cpp
@@ -295,7 +295,7 @@ bool ctlSQLBox::DoFind(const wxString &find, const wxString &replace, bool doRep
         selEnd = selStart + find.Length();
     }
 
-    if (selStart >= 0 && selStart != (unsigned int)(-1))
+    if (selStart >= 0 && selStart != (size_t)(-1))
     {
         if (reverse)
         {
